Add instrumented integration tests to rest module
We want to test the entire rest module as one unit (without adding it to other modules). So inside the rest module:
Create custom test runner...
to use new mock application...
to use new mock dagger component...
to use new mock dagger module...
to provide new mock ApiConstants and mock Context

Because we are instrumentation testing an Android library module (not an application module) we won't get the app's manifest. So we also have to create our own that has internet permission.